cleanup -started simplifying the widget & removing unnecessary things
