+++ classes/actions/class-content-scan.php
@@ -173,8 +173,6 @@ class Content_Scan extends Content {
 		foreach ( $posts as $post ) {
 			// Set the activity.
 			$activities[ $post->ID ] = \progress_planner()->get_activities__content_helpers()->get_activity_from_post( $post );
-			// Set the word count.
-			\progress_planner()->get_activities__content_helpers()->get_word_count( $post->post_content, $post->ID );
 		}
 
 		\progress_planner()->get_activities__query()->insert_activities( $activities );
