+++ classes/admin/class-page.php
@@ -51,7 +51,7 @@ class Page {
 			\progress_planner()->get_admin__widgets__challenge(),
 			\progress_planner()->get_admin__widgets__latest_badge(),
 			\progress_planner()->get_admin__widgets__badge_streak(),
-			\progress_planner()->get_admin__widgets__published_content(),
+			\progress_planner()->get_admin__widgets__content_managed(),
 			\progress_planner()->get_admin__widgets__whats_new(),
 		];
